Update backend.h
I rewrote the isValidItem function, edited the addItemToCart function to have a default return statement, added a private data member for the subtotal of the cart, and edited the name of the isValidItem function to match the naming convention of the other functions. I would also recommend splitting declaration and implementation of the member functions into a header and .cpp file, minus the inline functions.